OneSentence_20190806

+++ OneSentence/Config.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
 using System.Drawing;
 using System.IO;
 using System.Text;
@@ -11,12 +8,13 @@ namespace OneSentence
 {
     public partial class Config : Form
     {
-        Main main;
+        private Main main;//存储主窗体
 
         public Config(Main main)
         {
             InitializeComponent();
             this.main = main;
+            GoToolTip();//提示显示增加当前状态
         }
 
         /// <summary>
@@ -25,6 +23,7 @@ namespace OneSentence
         private void But_TopMost_Click(object sender, EventArgs e)
         {
             main.TopMost = !main.TopMost;
+            GoToolTip();//提示显示增加当前状态
         }
 
         /// <summary>
@@ -34,14 +33,15 @@ namespace OneSentence
         {
             if (main.FormBorderStyle == System.Windows.Forms.FormBorderStyle.Sizable)
             {
-                main.FormBorderStyle = System.Windows.Forms.FormBorderStyle.None;
-                main.TransparencyKey = System.Drawing.Color.Gray;
-                main.BackColor = System.Drawing.Color.Gray;
+                main.Location = new Point(main.Location.X, main.Location.Y + 35);//保持字体位置不变
+                main.FormBorderStyle = System.Windows.Forms.FormBorderStyle.None;//设置窗体样式
+                //main.SetLayered(main.BackColor);//设置透明并且实现击穿效果
             }
             else
             {
-                main.FormBorderStyle = System.Windows.Forms.FormBorderStyle.Sizable;
-                main.TransparencyKey = System.Drawing.Color.Transparent;
+                main.Location = new Point(main.Location.X, main.Location.Y - 35);//保持字体位置不变
+                main.FormBorderStyle = System.Windows.Forms.FormBorderStyle.Sizable;//设置窗体样式
+                //main.TransparencyKey = System.Drawing.Color.Transparent;
             }
         }
 
@@ -65,18 +65,19 @@ namespace OneSentence
 
             this.Cd_Color.ShowDialog();
             main.Txt_hitokoto.ForeColor = this.Cd_Color.Color;
+            Color color = Color.FromArgb(
+                this.Cd_Color.Color.R + (this.Cd_Color.Color.R >= 250 ? -5 : this.Cd_Color.Color.R < 5 ? 5 : 0),
+                this.Cd_Color.Color.G + (this.Cd_Color.Color.G >= 250 ? -5 : this.Cd_Color.Color.G < 5 ? 5 : 0),
+                this.Cd_Color.Color.B + (this.Cd_Color.Color.B >= 250 ? -5 : this.Cd_Color.Color.B < 5 ? 5 : 0));
+            main.SetLayered(color);//设置透明并且实现击穿效果
         }
 
         /// <summary>
-        /// 调整背景颜色
+        /// 调整配置文件
         /// </summary>
-        private void But_BColor_Click(object sender, EventArgs e)
+        private void But_Config_Click(object sender, EventArgs e)
         {
-            this.Cd_Color.Color = main.BackColor;
-
-            this.Cd_Color.ShowDialog();
-            main.BackColor = this.Cd_Color.Color;
-            main.TransparencyKey = this.Cd_Color.Color;
+            System.Diagnostics.Process.Start("notepad.exe", "Config.ini");
         }
 
         /// <summary>
@@ -85,6 +86,8 @@ namespace OneSentence
         private void But_Auto_Click(object sender, EventArgs e)
         {
             main.Txt_hitokoto.AutoSize = !main.Txt_hitokoto.AutoSize;
+            main.Txt_hitokoto.Width = main.Width - 26;
+            GoToolTip();//提示显示增加当前状态
         }
 
         /// <summary>
@@ -101,10 +104,22 @@ namespace OneSentence
         private void But_customize_Click(object sender, EventArgs e)
         {
             string OneSentence = "";//默认输出为空文本
-            DialogResult DRB = MessageBox.Show("需要输出示例文件？", "提示", MessageBoxButtons.YesNo);
-            if (DRB == DialogResult.Yes) OneSentence = "忘羡一曲远，曲终人不散\r\n苹果是给那些为了爱选择死亡的人的奖励\r\n我从远方赶来，恰巧你们也在";
-            File.WriteAllText("OneSentence.ini", OneSentence, Encoding.UTF8);
+            if (!File.Exists("OneSentence.ini"))//判断是否存在该文件
+            {
+                DialogResult DRB = MessageBox.Show("需要输出示例文件？", "提示", MessageBoxButtons.YesNo);
+                if (DRB == DialogResult.Yes) OneSentence = "忘羡一曲远，曲终人不散\r\n苹果是给那些为了爱选择死亡的人的奖励\r\n我从远方赶来，恰巧你们也在";
+                File.WriteAllText("OneSentence.ini", OneSentence, Encoding.UTF8);
+            }
             System.Diagnostics.Process.Start("notepad.exe", "OneSentence.ini");
+        }
+
+        /// <summary>
+        /// 提示显示增加当前状态
+        /// </summary>
+        private void GoToolTip()
+        {
+            Tt_toolTip.SetToolTip(But_TopMost, "显示的字体是否在所有的窗体之上\n当前状态：" + (main.TopMost ? "打开置顶" : "关闭置顶"));
+            Tt_toolTip.SetToolTip(But_Auto, "句子自动换行\n当前状态：" + (main.Txt_hitokoto.AutoSize ? "关闭换行" : "打开换行"));
         }
     }
 }
